Resolve storage conflicts
