+++ src/routes/job-history.ts
@@ -4,6 +4,20 @@
  */
 
 import { ApplicantProfile, JobHistoryEntry, JobHistorySubmission, JobRating, JobHistoryRequest } from '../lib/types';
+import { 
+  getApplicantProfile, 
+  createApplicantProfile, 
+  updateApplicantProfile,
+  saveJobHistorySubmission,
+  updateJobHistorySubmission,
+  saveJobHistoryEntry,
+  getJobHistoryByApplicant,
+  getJobHistorySubmissions,
+  getJobById,
+  saveJobRating,
+  getJobRatingsByApplicant,
+  type StorageEnv
+} from '../lib/storage';
 
 export async function handleJobHistoryPost(request: Request, env: any): Promise<Response> {
   try {
@@ -18,10 +32,12 @@ export async function handleJobHistoryPost(request: Request, env: any): Promise<
       });
     }
 
+    const storageEnv = env as StorageEnv;
+
     // Get or create applicant profile
-    let applicant = await getApplicantProfile(env, body.user_id);
+    let applicant = await getApplicantProfile(storageEnv, body.user_id);
     if (!applicant) {
-      applicant = await createApplicantProfile(env, body.user_id);
+      applicant = await createApplicantProfile(storageEnv, body.user_id);
     }
 
     // Create job history submission record
@@ -35,7 +51,7 @@ export async function handleJobHistoryPost(request: Request, env: any): Promise<
       submitted_at: new Date().toISOString()
     };
 
-    await saveJobHistorySubmission(env, submission);
+    await saveJobHistorySubmission(storageEnv, submission);
 
     // Process the job history with AI
     try {
@@ -45,17 +61,17 @@ export async function handleJobHistoryPost(request: Request, env: any): Promise<
       const savedEntries = [];
       for (const entry of processedHistory.entries) {
         entry.applicant_id = applicant.id!;
-        const savedEntry = await saveJobHistoryEntry(env, entry);
+        const savedEntry = await saveJobHistoryEntry(storageEnv, entry);
         savedEntries.push(savedEntry);
       }
 
       // Update applicant profile with extracted information
       if (processedHistory.profile_updates) {
-        await updateApplicantProfile(env, applicant.id!, processedHistory.profile_updates);
+        await updateApplicantProfile(storageEnv, applicant.id!, processedHistory.profile_updates);
       }
 
       // Update submission status
-      await updateJobHistorySubmission(env, submissionId, {
+      await updateJobHistorySubmission(storageEnv, submissionId, {
         processing_status: 'completed',
         ai_response: JSON.stringify(processedHistory),
         processed_entries: savedEntries.length,
@@ -75,7 +91,7 @@ export async function handleJobHistoryPost(request: Request, env: any): Promise<
 
     } catch (aiError) {
       // Update submission with error
-      await updateJobHistorySubmission(env, submissionId, {
+      await updateJobHistorySubmission(storageEnv, submissionId, {
         processing_status: 'failed',
         processing_error: aiError instanceof Error ? aiError.message : 'AI processing failed',
         processed_at: new Date().toISOString()
@@ -113,7 +129,8 @@ export async function handleJobHistoryGet(request: Request, env: any, params: Re
       });
     }
 
-    const applicant = await getApplicantProfile(env, user_id);
+    const storageEnv = env as StorageEnv;
+    const applicant = await getApplicantProfile(storageEnv, user_id);
     if (!applicant) {
       return new Response(JSON.stringify({ error: 'Applicant profile not found' }), { 
         status: 404,
@@ -121,8 +138,8 @@ export async function handleJobHistoryGet(request: Request, env: any, params: Re
       });
     }
 
-    const jobHistory = await getJobHistoryByApplicant(env, applicant.id!);
-    const submissions = await getJobHistorySubmissions(env, applicant.id!);
+    const jobHistory = await getJobHistoryByApplicant(storageEnv, applicant.id!);
+    const submissions = await getJobHistorySubmissions(storageEnv, applicant.id!);
 
     return new Response(JSON.stringify({
       applicant: applicant,
@@ -157,7 +174,8 @@ export async function handleJobRatingPost(request: Request, env: any): Promise<R
       });
     }
 
-    const applicant = await getApplicantProfile(env, body.user_id);
+    const storageEnv = env as StorageEnv;
+    const applicant = await getApplicantProfile(storageEnv, body.user_id);
     if (!applicant) {
       return new Response(JSON.stringify({ error: 'Applicant profile not found' }), { 
         status: 404,
@@ -166,7 +184,7 @@ export async function handleJobRatingPost(request: Request, env: any): Promise<R
     }
 
     // Get job details
-    const job = await getJobById(env, body.job_id);
+    const job = await getJobById(storageEnv, body.job_id);
     if (!job) {
       return new Response(JSON.stringify({ error: 'Job not found' }), { 
         status: 404,
@@ -175,7 +193,7 @@ export async function handleJobRatingPost(request: Request, env: any): Promise<R
     }
 
     // Get applicant's job history
-    const jobHistory = await getJobHistoryByApplicant(env, applicant.id!);
+    const jobHistory = await getJobHistoryByApplicant(storageEnv, applicant.id!);
 
     // Generate job rating using AI
     const rating = await generateJobRating(env, applicant, jobHistory, job);
@@ -183,7 +201,7 @@ export async function handleJobRatingPost(request: Request, env: any): Promise<R
     rating.job_id = body.job_id;
 
     // Save or update the rating
-    const savedRating = await saveJobRating(env, rating);
+    const savedRating = await saveJobRating(storageEnv, rating);
 
     return new Response(JSON.stringify(savedRating), {
       status: 200,
@@ -212,7 +230,8 @@ export async function handleJobRatingsGet(request: Request, env: any, params: Re
       });
     }
 
-    const applicant = await getApplicantProfile(env, user_id);
+    const storageEnv = env as StorageEnv;
+    const applicant = await getApplicantProfile(storageEnv, user_id);
     if (!applicant) {
       return new Response(JSON.stringify({ error: 'Applicant profile not found' }), { 
         status: 404,
@@ -220,7 +239,7 @@ export async function handleJobRatingsGet(request: Request, env: any, params: Re
       });
     }
 
-    const ratings = await getJobRatingsByApplicant(env, applicant.id!);
+    const ratings = await getJobRatingsByApplicant(storageEnv, applicant.id!);
 
     return new Response(JSON.stringify(ratings), {
       status: 200,
@@ -236,197 +255,6 @@ export async function handleJobRatingsGet(request: Request, env: any, params: Re
       headers: { 'Content-Type': 'application/json' }
     });
   }
-}
-
-// Helper functions for database operations
-
-async function getApplicantProfile(env: any, userId: string): Promise<ApplicantProfile | null> {
-  const result = await env.DB.prepare(
-    'SELECT * FROM applicant_profiles WHERE user_id = ?'
-  ).bind(userId).first();
-  
-  if (!result) return null;
-  
-  return {
-    ...result,
-    target_roles: result.target_roles ? JSON.parse(result.target_roles) : [],
-    skills: result.skills ? JSON.parse(result.skills) : [],
-    preferences: result.preferences ? JSON.parse(result.preferences) : {}
-  } as ApplicantProfile;
-}
-
-async function createApplicantProfile(env: any, userId: string): Promise<ApplicantProfile> {
-  const id = crypto.randomUUID();
-  const profile: ApplicantProfile = {
-    id,
-    user_id: userId,
-    created_at: new Date().toISOString(),
-    updated_at: new Date().toISOString()
-  };
-
-  await env.DB.prepare(`
-    INSERT INTO applicant_profiles (id, user_id, created_at, updated_at)
-    VALUES (?, ?, ?, ?)
-  `).bind(id, userId, profile.created_at, profile.updated_at).run();
-
-  return profile;
-}
-
-async function updateApplicantProfile(env: any, applicantId: string, updates: Partial<ApplicantProfile>): Promise<void> {
-  const setClause = [];
-  const values = [];
-  
-  for (const [key, value] of Object.entries(updates)) {
-    if (value !== undefined && key !== 'id' && key !== 'user_id') {
-      setClause.push(`${key} = ?`);
-      if (Array.isArray(value) || typeof value === 'object') {
-        values.push(JSON.stringify(value));
-      } else {
-        values.push(value);
-      }
-    }
-  }
-  
-  if (setClause.length === 0) return;
-  
-  setClause.push('updated_at = ?');
-  values.push(new Date().toISOString());
-  values.push(applicantId);
-
-  await env.DB.prepare(`
-    UPDATE applicant_profiles 
-    SET ${setClause.join(', ')} 
-    WHERE id = ?
-  `).bind(...values).run();
-}
-
-async function saveJobHistorySubmission(env: any, submission: JobHistorySubmission): Promise<void> {
-  await env.DB.prepare(`
-    INSERT INTO job_history_submissions 
-    (id, applicant_id, raw_content, content_type, processing_status, submitted_at)
-    VALUES (?, ?, ?, ?, ?, ?)
-  `).bind(
-    submission.id,
-    submission.applicant_id,
-    submission.raw_content,
-    submission.content_type,
-    submission.processing_status,
-    submission.submitted_at
-  ).run();
-}
-
-async function updateJobHistorySubmission(env: any, submissionId: string, updates: Partial<JobHistorySubmission>): Promise<void> {
-  const setClause = [];
-  const values = [];
-  
-  for (const [key, value] of Object.entries(updates)) {
-    if (value !== undefined && key !== 'id') {
-      setClause.push(`${key} = ?`);
-      values.push(value);
-    }
-  }
-  
-  if (setClause.length === 0) return;
-  
-  values.push(submissionId);
-
-  await env.DB.prepare(`
-    UPDATE job_history_submissions 
-    SET ${setClause.join(', ')} 
-    WHERE id = ?
-  `).bind(...values).run();
-}
-
-async function saveJobHistoryEntry(env: any, entry: JobHistoryEntry): Promise<JobHistoryEntry> {
-  const id = entry.id || crypto.randomUUID();
-  const now = new Date().toISOString();
-  
-  await env.DB.prepare(`
-    INSERT INTO job_history 
-    (id, applicant_id, company_name, job_title, department, employment_type, 
-     start_date, end_date, is_current, location, salary_min, salary_max, 
-     salary_currency, responsibilities, achievements, skills_used, technologies, 
-     keywords, created_at, updated_at)
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-  `).bind(
-    id, entry.applicant_id, entry.company_name, entry.job_title, entry.department,
-    entry.employment_type, entry.start_date, entry.end_date, entry.is_current ? 1 : 0,
-    entry.location, entry.salary_min, entry.salary_max, entry.salary_currency,
-    entry.responsibilities, entry.achievements, 
-    entry.skills_used ? JSON.stringify(entry.skills_used) : null,
-    entry.technologies ? JSON.stringify(entry.technologies) : null,
-    entry.keywords ? JSON.stringify(entry.keywords) : null,
-    now, now
-  ).run();
-
-  return { ...entry, id, created_at: now, updated_at: now };
-}
-
-async function getJobHistoryByApplicant(env: any, applicantId: string): Promise<JobHistoryEntry[]> {
-  const results = await env.DB.prepare(
-    'SELECT * FROM job_history WHERE applicant_id = ? ORDER BY start_date DESC'
-  ).bind(applicantId).all();
-  
-  return results.results.map((row: any) => ({
-    ...row,
-    is_current: Boolean(row.is_current),
-    skills_used: row.skills_used ? JSON.parse(row.skills_used) : [],
-    technologies: row.technologies ? JSON.parse(row.technologies) : [],
-    keywords: row.keywords ? JSON.parse(row.keywords) : []
-  })) as JobHistoryEntry[];
-}
-
-async function getJobHistorySubmissions(env: any, applicantId: string): Promise<JobHistorySubmission[]> {
-  const results = await env.DB.prepare(
-    'SELECT * FROM job_history_submissions WHERE applicant_id = ? ORDER BY submitted_at DESC'
-  ).bind(applicantId).all();
-  
-  return results.results as JobHistorySubmission[];
-}
-
-async function getJobById(env: any, jobId: string): Promise<any> {
-  return await env.DB.prepare('SELECT * FROM jobs WHERE id = ?').bind(jobId).first();
-}
-
-async function saveJobRating(env: any, rating: JobRating): Promise<JobRating> {
-  const id = rating.id || crypto.randomUUID();
-  const now = new Date().toISOString();
-  
-  // Use INSERT OR REPLACE to handle duplicates
-  await env.DB.prepare(`
-    INSERT OR REPLACE INTO job_ratings 
-    (id, applicant_id, job_id, overall_score, skill_match_score, experience_match_score,
-     compensation_fit_score, location_fit_score, company_culture_score, growth_potential_score,
-     rating_summary, recommendation, strengths, gaps, improvement_suggestions,
-     created_at, updated_at)
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-  `).bind(
-    id, rating.applicant_id, rating.job_id, rating.overall_score, rating.skill_match_score,
-    rating.experience_match_score, rating.compensation_fit_score, rating.location_fit_score,
-    rating.company_culture_score, rating.growth_potential_score, rating.rating_summary,
-    rating.recommendation, 
-    rating.strengths ? JSON.stringify(rating.strengths) : null,
-    rating.gaps ? JSON.stringify(rating.gaps) : null,
-    rating.improvement_suggestions, now, now
-  ).run();
-
-  return { ...rating, id, created_at: now, updated_at: now };
-}
-
-async function getJobRatingsByApplicant(env: any, applicantId: string): Promise<JobRating[]> {
-  const results = await env.DB.prepare(`
-    SELECT jr.*, j.title as job_title, j.company as job_company, j.url as job_url
-    FROM job_ratings jr
-    JOIN jobs j ON jr.job_id = j.id
-    WHERE jr.applicant_id = ?
-    ORDER BY jr.overall_score DESC
-  `).bind(applicantId).all();
-  
-  return results.results.map((row: any) => ({
-    ...row,
-    strengths: row.strengths ? JSON.parse(row.strengths) : [],
-    gaps: row.gaps ? JSON.parse(row.gaps) : []
-  })) as JobRating[];
 }
 
 // AI Processing Functions
